requirements.txt
# Core de Strands - Framework principal
strands-agents==0.2.1
strands-agents-tools==0.1.8
mcp==1.10.1

# Modelos de IA
ollama==0.5.1

# AWS SDK (para aws-bedrock-agent.py)
boto3==1.39.3
aws-requests-auth==0.4.3

# Utilidades esenciales
rich==14.0.0
dill==0.4.0
python-dotenv==1.1.1
fastapi==0.115.6
